requirements.txt
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
scikit-learn
tqdm
